enable associate to  branches on Crud organic_unit

+++ app/Livewire/OrganicUnit/Index.php
@@ -4,6 +4,7 @@ namespace App\Livewire\OrganicUnit;
 
 
 use App\Models\OrganicUnit;
+use App\Models\Branch;
 use Livewire\Component;
 use Livewire\Attributes\Rule;
 use TallStackUi\Traits\Interactions;
@@ -22,23 +23,50 @@ class Index extends Component
 
     // #[Rule('required', as: '"Nome"')]
     public $name;
+	
+	public $rows 			= [];
+	public $branch_id 		= [];
+	
+	public $showForm = 0;
 
     protected function rules()
     {
         $rules = [
-            'name' => 'required|unique:branches,name,' . $this->id,
+            'name' => 'required|unique:organic_units,name,' . $this->id,
         ];
 
         return $rules;
     }
+	
+    public function addRow(){
+        $this->rows[] = [
+            'branch_id' 		=> '',
+        ];
+    }
+    public function removeRow($index)
+    {
+        unset($this->rows[$index]);
+        $this->rows = array_values($this->rows);
+    }
+
+	public function create()
+    {
+        $this->reset();
+        $this->showForm	= true;
+
+    }	
 
     public function store()
     {
         $validated = $this->validate();
-        OrganicUnit::create($validated);
+        $query = OrganicUnit::create($validated);
+		
+        foreach ($this->rows as $value) {
+            $query->branches()->attach($value['branch_id']);
+        }		
         $this->reset();
         $this->resetValidation();
-		$this->dialog()->success('Successo', 'Adicionado com Sucesso.');
+		$this->dialog()->success('Successo', 'Adicionado com Sucesso.')->send();
     }
 
     public function edit($id)
@@ -47,6 +75,16 @@ class Index extends Component
         $query                          = OrganicUnit::findOrFail($id);
         $this->id                       = $id;
         $this->name                     = $query->name;
+		
+        if($query) {
+            foreach ($query->branches as $value){
+                $this->rows[] = [
+                    'branch_id'       => $value->pivot->branch_id,
+                ];
+
+            }
+        }
+        $this->showForm                 = true;				
     }
 
     public function update()
@@ -56,37 +94,38 @@ class Index extends Component
         {
             $query = OrganicUnit::findOrFail($this->id);
             $query->update($validated);
+			
+            $query->branches()->detach();
+
+            foreach ($this->rows as $value) {
+                $query->branches()->attach($value['toll_price_id']);
+            }			
             $this->reset();
             $this->resetValidation();
-			$this->dialog()->success('Successo', 'Editado com Sucesso.');
+			$this->dialog()->success('Successo', 'Editado com Sucesso.')->send();
         }
     }
 
     public function deleteConfirm($id)
     {
         $this->id = $id;
-        $this->dialog()->confirm('Atenção!', 'Tem certeza que deseja eliminar?', [
-            'confirm' => [
-                'text' => 'Confirmar',
-                'method' => 'delete',
-                // 'params' => 'Confirmed Successfully' // Can be a string or array
-            ],
-            /* Cancel is optional */
-            'cancel' => [
-                'text' => 'Cancelar',
-                'method' => 'cancel',
-                // 'params' => 'Cancelled Successfully' // Can be a string or array
-            ]
-        ]);
+		$this->dialog()
+			->question('Atenção!', 'Tem certeza que deseja eliminar?')
+			->confirm('Confirmar', 'delete')
+			->cancel('Cancelar', 'cancel')
+			->send();	
     }
 
     public function delete()
     {
         $query = OrganicUnit::where('id',$this->id)->first();
+		if ($query && $query->branches()->exists()) {
+           return $this->dialog()->error('Erro ao Eliminar', 'Não pode ser eliminado pois está associado a outro registo.')->send();
+        }
         $query->delete();
         $this->reset();
         $this->resetValidation();
-		$this->dialog()->success('Successo', 'Eliminado com Sucesso.');
+		$this->dialog()->success('Successo', 'Eliminado com Sucesso.')->send();
     }
     public function cancel()
     {
@@ -97,6 +136,7 @@ class Index extends Component
     public function render()
     {
         $query = OrganicUnit::all();
-        return view('livewire.organic-unit.index', compact('query'));
+		$query2 = Branch::all();
+        return view('livewire.organic-unit.index', compact('query','query2'));
     }
 }
